+++ __tests__/index.test.ts
@@ -1,5 +1,7 @@
 import SimpleValidationTests from './simple-validation'
 import NestedValidationTests from './nested-validation'
+import PrimitiveValidator from "./primitive-validator";
 
 describe('Simple Validation Tests', SimpleValidationTests)
 describe('Nested Validation Tests', NestedValidationTests)
+describe('Primitive Validation Tests', PrimitiveValidator)