+++ src/validators/main/validator.ts
@@ -1,6 +1,6 @@
 import {
   BshBatchValidationError,
-  BshValidationError,
+  BshValidationError, ValidatorComplexResultObjects,
   ValidatorComplexResultVFieldName,
   ValidatorResult,
   ValidatorResultArrays,
@@ -12,7 +12,6 @@ import {
   BatchValidatorResultInfo,
   ExtractNonPrimitiveKeys,
   NestedType,
-  ValidatorComplexResultObjects,
   ValidatorConfig,
   ValidatorOptions,
   ValidatorResultInfo,